Jem/72 doubles in ledger response (#73)
* [#72] Model amount types in ledger response as NativeAmount

* [#72] Correct display of Amount units

* correctly display amount

+++ src/main/scala/stellar/sdk/model/OrderBook.scala
@@ -33,7 +33,7 @@ object OrderBookDeserializer extends ResponseParser[OrderBook]({ o: JObject =>
           n = (c \ "price_r" \ "n").extract[Int],
           d = (c \ "price_r" \ "d").extract[Int]
         ),
-        quantity = Amount.toBaseUnits((c \ "amount").extract[String].toDouble).get
+        quantity = Amount.toBaseUnits((c \ "amount").extract[String]).get
       ))
   }
